pip_requirements.txt
numpy==1.20.3
jsonlines==2.0.0
pytorch-lightning==1.4.9
pandas==1.3.2
pywikibot==6.5.0
wikitextparser==0.47.5
wandb==0.12.2
transformers==4.10.0
omegaconf==2.1.1
hydra-core==1.1.1
hydra-colorlog==1.1.0
rich==10.10.0
pre-commit==2.15.0
matplotlib==3.4.2
tqdm==4.62.2
jupyter==1.0.0
ipykernel==6.4.1
jupyterlab==3.1.7
seaborn==0.11.2
